+++ src/Classes/Validation.php
@@ -2,12 +2,22 @@
 
 namespace AshAllenDesign\LaravelExchangeRates\Classes;
 
+use AshAllenDesign\LaravelExchangeRates\Exceptions\ExchangeRateException;
 use AshAllenDesign\LaravelExchangeRates\Exceptions\InvalidCurrencyException;
 use AshAllenDesign\LaravelExchangeRates\Exceptions\InvalidDateException;
 use Carbon\Carbon;
 
 class Validation
 {
+    /**
+     * A Carbon object for the earliest possible date that
+     * an exchange rate can be fetched for. The date is:
+     * 4th January 2020.
+     *
+     * @var Carbon|null
+     */
+    private static $earliestPossibleDate;
+
     /**
      * Validate that the currency is supported by the
      * Exchange Rates API.
@@ -16,12 +26,30 @@ class Validation
      *
      * @throws InvalidCurrencyException
      */
-    public static function validateCurrencyCode(string $currencyCode)
+    public static function validateCurrencyCode(string $currencyCode): void
     {
         $currencies = new Currency();
 
         if (! $currencies->isAllowableCurrency($currencyCode)) {
             throw new InvalidCurrencyException($currencyCode.' is not a valid country code.');
+        }
+    }
+
+    /**
+     * Validate that the currencies are all supported by
+     * the Exchange Rates API.
+     *
+     * @param  array  $currencyCodes
+     * @throws InvalidCurrencyException
+     */
+    public static function validateCurrencyCodes(array $currencyCodes): void
+    {
+        $currencies = new Currency();
+
+        foreach ($currencyCodes as $currencyCode) {
+            if (! $currencies->isAllowableCurrency($currencyCode)) {
+                throw new InvalidCurrencyException($currencyCode.' is not a valid country code.');
+            }
         }
     }
 
@@ -35,7 +63,7 @@ class Validation
      *
      * @throws InvalidDateException
      */
-    public static function validateStartAndEndDates(Carbon $from, Carbon $to)
+    public static function validateStartAndEndDates(Carbon $from, Carbon $to): void
     {
         self::validateDate($from);
         self::validateDate($to);
@@ -53,10 +81,31 @@ class Validation
      *
      * @throws InvalidDateException
      */
-    public static function validateDate(Carbon $date)
+    public static function validateDate(Carbon $date): void
     {
         if (! $date->isPast()) {
             throw new InvalidDateException('The date must be in the past.');
+        }
+
+        if (! self::$earliestPossibleDate) {
+            self::$earliestPossibleDate = Carbon::createFromDate(1999, 1, 4)->startOfDay();
+        }
+
+        if ($date->isBefore(static::$earliestPossibleDate)) {
+            throw new InvalidDateException('The date cannot be before 4th January 1999.');
+        }
+    }
+
+    /**
+     * Validate that the parameter is a string or array.
+     *
+     * @param $paramToValidate
+     * @throws ExchangeRateException
+     */
+    public static function validateIsStringOrArray($paramToValidate): void
+    {
+        if (! is_string($paramToValidate) && ! is_array($paramToValidate)) {
+            throw new ExchangeRateException($paramToValidate.' is not a string or array.');
         }
     }
 }