+++ src/Exceptions/ExchangeRateException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace AshAllenDesign\LaravelExchangeRates\Exceptions;
+
+use Exception;
+
+class ExchangeRateException extends Exception
+{
+}