+++ tests/Unit/ConvertBetweenDateRangeTest.php
@@ -4,6 +4,7 @@ namespace AshAllenDesign\LaravelExchangeRates\Tests\Unit;
 
 use AshAllenDesign\LaravelExchangeRates\Classes\ExchangeRate;
 use AshAllenDesign\LaravelExchangeRates\Classes\RequestBuilder;
+use AshAllenDesign\LaravelExchangeRates\Exceptions\ExchangeRateException;
 use AshAllenDesign\LaravelExchangeRates\Exceptions\InvalidCurrencyException;
 use AshAllenDesign\LaravelExchangeRates\Exceptions\InvalidDateException;
 use Carbon\Carbon;
@@ -30,7 +31,7 @@ class ConvertBetweenDateRangeTest extends TestCase
                 ],
             ])
             ->once()
-            ->andReturn($this->mockResponse());
+            ->andReturn($this->mockResponseForOneSymbol());
 
         $exchangeRate = new ExchangeRate($requestBuilderMock);
         $currencies = $exchangeRate->convertBetweenDateRange(100, 'GBP', 'EUR', $fromDate, $toDate);
@@ -116,7 +117,7 @@ class ConvertBetweenDateRangeTest extends TestCase
                 ],
             ])
             ->once()
-            ->andReturn($this->mockResponse());
+            ->andReturn($this->mockResponseForOneSymbol());
 
         $exchangeRate = new ExchangeRate($requestBuilderMock);
         $currencies = $exchangeRate->shouldBustCache()->convertBetweenDateRange(100, 'GBP', 'EUR', $fromDate, $toDate);
@@ -138,6 +139,50 @@ class ConvertBetweenDateRangeTest extends TestCase
         ];
         $this->assertEquals($cachedExchangeRates,
             Cache::get('laravel_xr_GBP_EUR_'.$fromDate->format('Y-m-d').'_'.$toDate->format('Y-m-d')));
+    }
+
+    /** @test */
+    public function converted_values_can_be_returned_for_multiple_currencies()
+    {
+        $fromDate = now()->subWeek();
+        $toDate = now();
+
+        $requestBuilderMock = Mockery::mock(RequestBuilder::class)->makePartial();
+        $requestBuilderMock->expects('makeRequest')
+            ->withArgs([
+                '/history',
+                [
+                    'base'     => 'GBP',
+                    'start_at' => $fromDate->format('Y-m-d'),
+                    'end_at'   => $toDate->format('Y-m-d'),
+                    'symbols'  => 'EUR,USD',
+                ],
+            ])
+            ->once()
+            ->andReturn($this->mockResponseForMultipleSymbols());
+
+        $exchangeRate = new ExchangeRate($requestBuilderMock);
+        $currencies = $exchangeRate->convertBetweenDateRange(100, 'GBP', ['EUR', 'USD'], $fromDate, $toDate);
+
+        $expectedArray = [
+            '2019-11-08' => ['EUR' => 116.06583254, 'USD' => 111.11111111],
+            '2019-11-06' => ['EUR' => 116.23446817, 'USD' => 122.22222222],
+            '2019-11-07' => ['EUR' => 115.68450522, 'USD' => 133.33333333],
+            '2019-11-05' => ['EUR' => 116.12648497, 'USD' => 144.44444444],
+            '2019-11-04' => ['EUR' => 115.78362356, 'USD' => 155.55555555],
+        ];
+
+        $this->assertEquals($expectedArray, $currencies);
+
+        $cachedExchangeRates = [
+            '2019-11-08' => ['EUR' => 1.1606583254, 'USD' => 1.1111111111],
+            '2019-11-06' => ['EUR' => 1.1623446817, 'USD' => 1.2222222222],
+            '2019-11-07' => ['EUR' => 1.1568450522, 'USD' => 1.3333333333],
+            '2019-11-05' => ['EUR' => 1.1612648497, 'USD' => 1.4444444444],
+            '2019-11-04' => ['EUR' => 1.1578362356, 'USD' => 1.5555555555],
+        ];
+        $this->assertEquals($cachedExchangeRates,
+            Cache::get('laravel_xr_GBP_EUR_USD_'.$fromDate->format('Y-m-d').'_'.$toDate->format('Y-m-d')));
     }
 
     /** @test */
@@ -208,7 +253,7 @@ class ConvertBetweenDateRangeTest extends TestCase
         $this->expectExceptionMessage('INVALID is not a valid country code.');
 
         $exchangeRate = new ExchangeRate();
-        $exchangeRate->convertBetweenDateRange(100, 'GBP', 'INVALID', now()->subWeek(), now()->subDay());
+        $exchangeRate->convertBetweenDateRange(100, 'INVALID', 'GBP', now()->subWeek(), now()->subDay());
     }
 
     /** @test */
@@ -218,10 +263,20 @@ class ConvertBetweenDateRangeTest extends TestCase
         $this->expectExceptionMessage('INVALID is not a valid country code.');
 
         $exchangeRate = new ExchangeRate();
-        $exchangeRate->convertBetweenDateRange(100, 'INVALID', 'GBP', now()->subWeek(), now()->subDay());
+        $exchangeRate->convertBetweenDateRange(100, 'GBP', 'INVALID', now()->subWeek(), now()->subDay());
     }
 
-    private function mockResponse()
+    /** @test */
+    public function exception_is_thrown_if_the_to_parameter_is_not_a_string_or_array()
+    {
+        $this->expectException(ExchangeRateException::class);
+        $this->expectExceptionMessage('123 is not a string or array.');
+
+        $exchangeRate = new ExchangeRate();
+        $exchangeRate->convertBetweenDateRange(100, 'GBP', 123, now()->subWeek(), now()->subDay());
+    }
+
+    private function mockResponseForOneSymbol()
     {
         return [
             'rates'    => [
@@ -239,6 +294,37 @@ class ConvertBetweenDateRangeTest extends TestCase
                 ],
                 '2019-11-04' => [
                     'EUR' => 1.1578362356,
+                ],
+            ],
+            'start_at' => '2019-11-03',
+            'base'     => 'GBP',
+            'end_at'   => '2019-11-10',
+        ];
+    }
+
+    private function mockResponseForMultipleSymbols()
+    {
+        return [
+            'rates'    => [
+                '2019-11-08' => [
+                    'EUR' => 1.1606583254,
+                    'USD' => 1.1111111111,
+                ],
+                '2019-11-06' => [
+                    'EUR' => 1.1623446817,
+                    'USD' => 1.2222222222,
+                ],
+                '2019-11-07' => [
+                    'EUR' => 1.1568450522,
+                    'USD' => 1.3333333333,
+                ],
+                '2019-11-05' => [
+                    'EUR' => 1.1612648497,
+                    'USD' => 1.4444444444,
+                ],
+                '2019-11-04' => [
+                    'EUR' => 1.1578362356,
+                    'USD' => 1.5555555555,
                 ],
             ],
             'start_at' => '2019-11-03',